504_final_project/app/requirements.txt
Flask==2.3.2
PyMySQL==1.1.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
